// src/types/Player.ts
export interface Player {
  id: number;
  PLAYER: string;
  TEAM: string;
  AGE: number;
  HEIGHT: number;
  WEIGHT: number;
  COLLEGE: string;
  COUNTRY: string;
  'DRAFT YEAR': number;
  'DRAFT ROUND': string;
  'DRAFT NUMBER': string;
  GP: number;
  PTS: number;
  REB: number;
  AST: number;
  NETRTG: number;
  'OREB%': number;
  'DREB%': number;
  'USG%': number;
  'TS%': number;
  'AST%': number;
  COUN_CODE: string;
  imageUrl: string;
}